requirements.txt
python-dotenv
azure-identity
azure-ai-projects
openai
gradio
requests